Dra ut Behandling ruter til egen komponent
