+++ src/frontend/sider/Fagsak/Behandling/BehandlingContainer.tsx
@@ -1,29 +1,16 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
-import { Route, Routes, useLocation } from 'react-router';
 import styled from 'styled-components';
 
 import { Alert } from '@navikt/ds-react';
 import { ABorderDivider } from '@navikt/ds-tokens/dist/tokens';
 import { RessursStatus } from '@navikt/familie-typer';
 
+import { BehandlingRouter } from './BehandlingRouter';
 import { useBehandlingContext } from './context/BehandlingContext';
 import Høyremeny from './Høyremeny/Høyremeny';
-import Behandlingsresultat from './sider/Behandlingsresultat/Behandlingsresultat';
-import RegistrerSøknad from './sider/RegistrerSøknad/RegistrerSøknad';
-import { SøknadProvider } from './sider/RegistrerSøknad/SøknadContext';
-import type { SideId } from './sider/sider';
-import { sider } from './sider/sider';
-import Simulering from './sider/Simulering/Simulering';
-import { SimuleringProvider } from './sider/Simulering/SimuleringContext';
-import { SammensattKontrollsakProvider } from './sider/Vedtak/SammensattKontrollsak/SammensattKontrollsakContext';
-import Vedtak from './sider/Vedtak/Vedtak';
-import Vilkårsvurdering from './sider/Vilkårsvurdering/Vilkårsvurdering';
-import { VilkårsvurderingProvider } from './sider/Vilkårsvurdering/VilkårsvurderingContext';
-import { TidslinjeProvider } from '../../../komponenter/Tidslinje/TidslinjeContext';
 import type { IMinimalFagsak } from '../../../typer/fagsak';
 import type { IPersonInfo } from '../../../typer/person';
-import { hentSideHref } from '../../../utils/miljø';
 import { Fagsaklinje } from '../Fagsaklinje/Fagsaklinje';
 import Venstremeny from './Venstremeny/Venstremeny';
 
@@ -55,17 +42,7 @@ interface Props {
 }
 
 const BehandlingContainer: React.FunctionComponent<Props> = ({ bruker, minimalFagsak }) => {
-    const location = useLocation();
-    const { åpenBehandling, leggTilBesøktSide } = useBehandlingContext();
-
-    const sidevisning = hentSideHref(location.pathname);
-    useEffect(() => {
-        if (sidevisning) {
-            leggTilBesøktSide(
-                Object.entries(sider).find(([_, side]) => side.href === sidevisning)?.[0] as SideId
-            );
-        }
-    }, [sidevisning]);
+    const { åpenBehandling } = useBehandlingContext();
 
     switch (åpenBehandling.status) {
         case RessursStatus.SUKSESS:
@@ -77,59 +54,10 @@ const BehandlingContainer: React.FunctionComponent<Props> = ({ bruker, minimalFa
                             <Venstremeny />
                         </VenstremenyContainer>
                         <HovedinnholdContainer>
-                            <Routes>
-                                <Route
-                                    path="/registrer-soknad"
-                                    element={
-                                        <SøknadProvider åpenBehandling={åpenBehandling.data}>
-                                            <RegistrerSøknad />
-                                        </SøknadProvider>
-                                    }
-                                />
-                                <Route
-                                    path="/vilkaarsvurdering"
-                                    element={
-                                        <VilkårsvurderingProvider
-                                            åpenBehandling={åpenBehandling.data}
-                                        >
-                                            <Vilkårsvurdering
-                                                åpenBehandling={åpenBehandling.data}
-                                            />
-                                        </VilkårsvurderingProvider>
-                                    }
-                                />
-                                <Route
-                                    path="/tilkjent-ytelse"
-                                    element={
-                                        <TidslinjeProvider>
-                                            <Behandlingsresultat
-                                                åpenBehandling={åpenBehandling.data}
-                                            />
-                                        </TidslinjeProvider>
-                                    }
-                                />
-                                <Route
-                                    path="/simulering"
-                                    element={
-                                        <SimuleringProvider åpenBehandling={åpenBehandling.data}>
-                                            <Simulering åpenBehandling={åpenBehandling.data} />
-                                        </SimuleringProvider>
-                                    }
-                                />
-                                <Route
-                                    path="/vedtak"
-                                    element={
-                                        <SammensattKontrollsakProvider
-                                            åpenBehandling={åpenBehandling.data}
-                                        >
-                                            <Vedtak
-                                                åpenBehandling={åpenBehandling.data}
-                                                bruker={bruker}
-                                            />
-                                        </SammensattKontrollsakProvider>
-                                    }
-                                />
-                            </Routes>
+                            <BehandlingRouter
+                                åpenBehandling={åpenBehandling.data}
+                                bruker={bruker}
+                            />
                         </HovedinnholdContainer>
                         <HøyremenyContainer>
                             <Høyremeny bruker={bruker} />
